requirements.txt
fastapi
uvicorn[standard]
python-multipart
pydantic
requests
httpx
openai
python-docx
pdfplumber
openpyxl
python-pptx
PyMuPDF
Pillow
paddleocr
opencv-python
numpy
matplotlib
reportlab
SQLAlchemy
celery
redis
minio
psycopg2-binary
grpcio
grpcio-tools
